+++ penny_university_frontend/src/components/modal/auth/password.tsx
@@ -8,7 +8,7 @@ import {
   Form, ModalHeader, ModalBody, ModalFooter, Button,
 } from 'reactstrap'
 import modalDispatch from '../dispatch'
-import {dispatchLogin, requestPasswordReset} from '../../../actions/user'
+import { dispatchLogin, requestPasswordReset } from '../../../actions/user'
 import { Input } from '../../fields'
 
 type AuthPasswordModalProps = {
