+++ users/serializers.py
@@ -1,7 +1,11 @@
 from django.conf import settings
 from django.contrib.auth import authenticate
+from django.contrib.auth.tokens import default_token_generator
 from django.contrib.auth.forms import PasswordResetForm
+from django.utils.encoding import force_bytes
+from django.utils.http import urlsafe_base64_encode
 from rest_framework import exceptions, serializers, status
+from common.utils import build_url
 from .models import User, SocialProfile
 
 
@@ -111,6 +115,15 @@ class CustomPasswordResetSerializer(serializers.Serializer):
     def save(self):
         request = self.context.get('request')
 
+        user = User.objects.get(email=self.reset_form.cleaned_data['email'])
+
+        reset_password_url = build_url(
+            settings.FRONT_END_HOST,
+            'reset-password',
+            token=default_token_generator.make_token(user),
+            uid=urlsafe_base64_encode(force_bytes(user.pk))
+        )
+
         opts = {
             'use_https': request.is_secure(),
             'from_email': '"Penny University" <reset@pennyuniversity.org>',
@@ -119,8 +132,7 @@ class CustomPasswordResetSerializer(serializers.Serializer):
             'html_email_template_name': 'users/reset_password_email.html',
             'request': request,
             'extra_email_context': {
-                'base': settings.FRONT_END_HOST,
-                'endpoint': 'reset-password',
+                'reset_password_url': reset_password_url
             }
         }
         self.reset_form.save(**opts)
