+++ users/templates/users/reset_password_email.html
@@ -1,5 +1,3 @@
-{% load urls %}
-
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional //EN"
         "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
 <html xmlns="http://www.w3.org/1999/xhtml" xmlns:v="urn:schemas-microsoft-com:vml"
@@ -242,7 +240,6 @@
                                                         <v:textbox inset="0,0,0,0">
                                                             <center style="color:#ffffff; font-family:Arial, sans-serif; font-size:16px">
                                         <![endif]-->
-                                        {% build_url base endpoint uid=uid token=token as reset_password_url %}
                                         <a href="{{ reset_password_url }}">
                                             <div style="text-decoration:none;display:inline-block;color:#ffffff;background-color:#008dd5;border-radius:4px;-webkit-border-radius:4px;-moz-border-radius:4px;width:auto; width:auto;;border-top:1px solid #008dd5;border-right:1px solid #008dd5;border-bottom:1px solid #008dd5;border-left:1px solid #008dd5;padding-top:5px;padding-bottom:5px;font-family:'Roboto Slab', Arial, 'Helvetica Neue', Helvetica, sans-serif;text-align:center;mso-border-alt:none;word-break:keep-all;">
                                                 <span style="padding-left:20px;padding-right:20px;font-size:16px;display:inline-block;"><span
