+++ users/templates/users/reset_password_email.txt
@@ -1,6 +1,4 @@
-{% load urls %}
 {% autoescape off %}
-{% build_url base endpoint uid=uid token=token as reset_password_url %}
 It looks like you requested a password reset.
 Click the link below to reset your password.
 {{ reset_password_url }}
