+++ users/templates/users/verify_email.txt
@@ -1,6 +1,4 @@
-{% load urls %}
 {% autoescape off %}
-{% build_url base endpoint email=email token=token as verification_url%}
 Welcome to Penny University!
 Thank you for signing up, {{ first_name }}. Before we get started, please verify your email by clicking this link:
 {{ verification_url }}
